requirements.txt
opencv-python>=4.5.0
albumentations>=1.3.0
fontTools>=4.12.1
lmdb>=0.98
pillow>=8.0.0
numpy>=1.19.0
scipy>=1.5.0
tenacity>=8.0.0
fire>=0.4.0
tqdm>=4.60.0
loguru>=0.6.0
cffi>=1.14.0
sphinx>=3.1.2
sphinx-rtd-theme>=0.5.0
recommonmark>=0.6.0
sphinx-markdown-tables>=0.0.3
dataclasses>=0.6
pytest>=6.0.0
typer>=0.4.0
streamlit>=1.0.0
pre-commit>=3.0.0
